Set aside OMP for a while, testing priors for innovation varianc
Either we get good oscillation

or we get nearly flat latent state, fs wandering, or we can get probability of nearly 1 near spikes and 0 everywhere else.

Propose to control this by setting prior parameter a of inverse gamma.

We choose median of prior to be something reasonable for bin size in millisecond range, and we just adjust the a to avoid overfit of just the spikes.